+++ src/lib/dao/UserDAO.java
@@ -2,6 +2,6 @@ package lib.dao;
 
 import lib.entities.User;
 
-public class UserDAO extends GenericDaoImpl<User, Long> {
+public class UserDAO extends GenericDaoImpl<User, Integer> {
 	private static final long serialVersionUID = 3233237838684697121L;
 }
